add test for distributed set_index
