+++ sparsity/test/test_dask_sparse_frame.py
@@ -334,3 +334,17 @@ def test_sdf_sort_index():
     res = dsf.compute()
     assert res.index.is_monotonic
     assert res.columns.tolist() == list('ABCD')
+
+
+def test_set_index(clickstream):
+    ddf = dd.from_pandas(clickstream, npartitions=10)
+    dsf = one_hot_encode(ddf,
+                         categories={'page_id': list('ABCDE'),
+                                     'other_categorical': list('FGHIJ')},
+                         order=['other_categorical', 'page_id'],
+                         index_col=['index', 'id'])
+    dense = dsf.compute().set_index(level=1).todense()
+    res = dsf.set_index(level=1).compute().todense()
+
+    pdt.assert_frame_equal(dense, res)
+
